Merge and Quick sort algorithms implementation
Call the "sortJobSalaries" function in main. It takes in a map<key: Job title, value: salary> and returns two sorted vectors (in a vector of vectors), one in ascending using merge sort and one in descending order using quick sort. Use these vectors to print the top 10 salaries and the lowest 10 salaries. (do not call the merge sort, quick sort, or any other helper functions as the "sortJobSalaries" function already does that)